+++ app/src/com/mapbar/adas/HomePage.java
@@ -35,6 +35,7 @@ import com.mapbar.hamster.BlueManager;
 import com.mapbar.hamster.OBDEvent;
 import com.mapbar.hamster.OBDStatusInfo;
 import com.mapbar.hamster.core.ProtocolUtils;
+import com.mapbar.hamster.log.FileLoggingTree;
 import com.mapbar.hamster.log.Log;
 import com.miyuan.obd.R;
 
@@ -504,7 +505,9 @@ public class HomePage extends AppBasePage implements View.OnClickListener, BleCa
             builder.addPart(MultipartBody.Part.createFormData("serialNumber", obdStatusInfo.getSn()))
                     .addPart(MultipartBody.Part.createFormData("type", "1"));
             for (File file : logs) {
-                builder.addFormDataPart("file", file.getName(), RequestBody.create(MediaType.parse("application/octet-stream"), file));
+                if (!file.getName().equals(FileLoggingTree.fileName)) {
+                    builder.addFormDataPart("file", file.getName(), RequestBody.create(MediaType.parse("application/octet-stream"), file));
+                }
             }
             Request request = new Request.Builder()
                     .url(URLUtils.UPDATE_ERROR_FILE)
@@ -531,7 +534,9 @@ public class HomePage extends AppBasePage implements View.OnClickListener, BleCa
                                 }
                             });
                             for (File delete : logs) {
-                                delete.delete();
+                                if (!delete.getName().equals(FileLoggingTree.fileName)) {
+                                    delete.delete();
+                                }
                             }
                         }
                     } catch (JSONException e) {
