+++ app/src/com/mapbar/adas/ProtocolCheckFailPage.java
@@ -17,6 +17,7 @@ import com.mapbar.hamster.BlueManager;
 import com.mapbar.hamster.OBDEvent;
 import com.mapbar.hamster.OBDStatusInfo;
 import com.mapbar.hamster.core.ProtocolUtils;
+import com.mapbar.hamster.log.FileLoggingTree;
 import com.mapbar.hamster.log.Log;
 import com.miyuan.obd.R;
 
@@ -284,6 +285,9 @@ public class ProtocolCheckFailPage extends AppBasePage implements BleCallBackLis
     }
 
     private void uploadLog() {
+        if (null == obdStatusInfo) {
+            return;
+        }
         Log.d("ProtocolCheckFailPage uploadLog ");
         final File dir = new File(Environment.getExternalStorageDirectory().getPath() + File.separator + "obd");
         final File[] logs = dir.listFiles();
@@ -293,7 +297,9 @@ public class ProtocolCheckFailPage extends AppBasePage implements BleCallBackLis
             builder.addPart(MultipartBody.Part.createFormData("serialNumber", obdStatusInfo.getSn()))
                     .addPart(MultipartBody.Part.createFormData("type", "1"));
             for (File file : logs) {
-                builder.addFormDataPart("file", file.getName(), RequestBody.create(MediaType.parse("application/octet-stream"), file));
+                if (!file.getName().equals(FileLoggingTree.fileName)) {
+                    builder.addFormDataPart("file", file.getName(), RequestBody.create(MediaType.parse("application/octet-stream"), file));
+                }
             }
             Request request = new Request.Builder()
                     .url(URLUtils.UPDATE_ERROR_FILE)
@@ -320,7 +326,9 @@ public class ProtocolCheckFailPage extends AppBasePage implements BleCallBackLis
                                 }
                             });
                             for (File delete : logs) {
-                                delete.delete();
+                                if (!delete.getName().equals(FileLoggingTree.fileName)) {
+                                    delete.delete();
+                                }
                             }
                         }
                     } catch (JSONException e) {
